requirements.txt
Django==3.2.13
paho-mqtt==1.6.1
